src/requirements.txt
appdirs==1.4.2
appnope==0.1.0
automium==0.2.6
automium-web==0.1.1
backports.shutil-get-terminal-size==1.0.0
click==6.7
decorator==4.0.11
Django==1.10.6
djangorestframework==3.5.4
enum34==1.1.6
Flask==0.12
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
legacy==0.1.6
MarkupSafe==0.23
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.11.15
wheel==0.24.0
